#42 Tests User Story 16.4 + quelques modifs 16.3

+++ src/main/java/fr/unice/polytech/kaleate/Supplement.java
@@ -2,6 +2,5 @@ package fr.unice.polytech.kaleate;
 
 public interface Supplement {
 
-    int getPrix();
-
+    float getPrix();
 }
