+++ src/main/java/fr/unice/polytech/kaleate/SupplementComposant.java
@@ -2,15 +2,15 @@ package fr.unice.polytech.kaleate;
 
 public class SupplementComposant extends Composant implements Supplement{
 
-    private int prix;
+    private float prix;
 
-    public SupplementComposant(String nomSupplementComposant, int prix) {
+    public SupplementComposant(String nomSupplementComposant, float prix) {
         super(nomSupplementComposant);
         this.prix = prix;
     }
 
     @Override
-    public int getPrix() {
+    public float getPrix() {
         return this.prix;
     }
 }
